--- code-samples/popular-topics/reactions/raw-event.js
@@ -1,42 +1,0 @@
-const Discord = require('discord.js');
-const client = new Discord.Client();
-
-client.once('ready', () => {
-	console.log('Ready!');
-});
-
-const events = {
-	MESSAGE_REACTION_ADD: 'messageReactionAdd',
-	MESSAGE_REACTION_REMOVE: 'messageReactionRemove',
-};
-
-client.on('raw', async event => {
-	if (!events.hasOwnProperty(event.t)) return;
-
-	const { d: data } = event;
-	const user = client.users.get(data.user_id);
-	const channel = client.channels.get(data.channel_id) || await user.createDM();
-
-	if (channel.messages.has(data.message_id)) return;
-
-	const message = await channel.fetchMessage(data.message_id);
-	const emojiKey = (data.emoji.id) ? `${data.emoji.name}:${data.emoji.id}` : data.emoji.name;
-	let reaction = message.reactions.get(emojiKey);
-
-	if (!reaction) {
-		const emoji = new Discord.Emoji(client.guilds.get(data.guild_id), data.emoji);
-		reaction = new Discord.MessageReaction(message, emoji, 1, data.user_id === client.user.id);
-	}
-
-	client.emit(events[event.t], reaction, user);
-});
-
-client.on('messageReactionAdd', (reaction, user) => {
-	console.log(`${user.username} reacted with "${reaction.emoji.name}".`);
-});
-
-client.on('messageReactionRemove', (reaction, user) => {
-	console.log(`${user.username} removed their "${reaction.emoji.name}" reaction.`);
-});
-
-client.login('your-token-goes-here');
