+++ code-samples/popular-topics/reactions/uncached-messages.js
@@ -1,0 +1,35 @@
+const Discord = require('discord.js');
+const client = new Discord.Client({ partials: ['MESSAGE', 'CHANNEL'] });
+
+client.once('ready', () => {
+	console.log('Ready!');
+});
+
+const events = {
+	MESSAGE_REACTION_ADD: 'messageReactionAdd',
+	MESSAGE_REACTION_REMOVE: 'messageReactionRemove',
+};
+
+client.on('messageReactionAdd', async (reaction, user) => {
+	if (reaction.message.partial) {
+		try {
+			await reaction.message.fetch();
+		} catch (error) {
+			console.log('Something went wrong when fetching the message: ', error);
+		}
+	}
+	console.log(`${user.username} reacted with "${reaction.emoji.name}".`);
+});
+
+client.on('messageReactionRemove', async (reaction, user) => {
+	if (reaction.message.partial) {
+		try {
+			await reaction.message.fetch();
+		} catch (error) {
+			console.log('Something went wrong when fetching the message: ', error);
+		}
+	}
+	console.log(`${user.username} removed their "${reaction.emoji.name}" reaction.`);
+});
+
+client.login('your-token-goes-here');
